--- src/components/Calendar/index.js
@@ -1,2 +1,0 @@
-export { default as Calendar } from './Calendar';
-export { default as Tile } from './Tile/Tile';
