+++ src/constants/index.js
@@ -1,10 +1,8 @@
 export * from './about.js';
 export * from './admin.js';
 export * from './align.js';
-export * from './attendance.js';
 export * from './dropDown.js';
 export * from './examReview.js';
-export * from './point.js';
 export * from './postCategories.js';
 export * from './search.js';
 export * from './seconds.js';
