+++ src/feature/attendance/component/index.js
@@ -1,0 +1,1 @@
+export { default as Calendar } from './Calendar/Calendar';
