+++ src/feature/attendance/constant/index.js
@@ -1,0 +1,2 @@
+export * from './attendance.js';
+export * from './point.js';
