+++ src/feature/attendance/constant/point.js
